[ feat ] 식당 권한 세부 설정 ( #5 )

+++ src/main/java/com/qring/restaurant/application/v1/service/RestaurantService.java
@@ -85,10 +85,18 @@ public class RestaurantService {
     @Transactional
     public void putBy(String passport, Long id, PutRestaurantReqDTOV1 dto) {
 
+        String userRole = PassportUtil.getRole(passport);
+        Long userId = PassportUtil.getUserId(passport);
+
         validateUserRole(PassportUtil.getRole(passport), "관리자", "점주");
 
         RestaurantEntity existingRestaurant = restaurantRepository.findByIdAndDeletedAtIsNull(id)
                 .orElseThrow(() -> new EntityNotFoundException("식당을 찾을 수 없습니다."));
+
+        // 점주일 경우 본인의 식당인지 확인
+        if (userRole.equals("점주") && !existingRestaurant.getUserId().equals(userId)) {
+            throw new UnauthorizedAccessException("본인의 식당만 수정할 수 있습니다.");
+        }
 
         CategoryEntity category = categoryRepository.findByIdAndDeletedAtIsNull(dto.getRestaurant().getCategoryId())
                 .orElseThrow(() -> new EntityNotFoundException("카테고리를 찾을 수 없습니다."));
@@ -122,10 +130,19 @@ public class RestaurantService {
     @Transactional
     public void deleteBy(String passport, Long id) {
 
+        String userRole = PassportUtil.getRole(passport);
+        Long userId = PassportUtil.getUserId(passport);
+
         validateUserRole(PassportUtil.getRole(passport), "관리자", "점주");
 
         RestaurantEntity restaurant = restaurantRepository.findByIdAndDeletedAtIsNull(id)
                 .orElseThrow(() -> new EntityNotFoundException("식당을 찾을 수 없습니다."));
+
+        // 점주일 경우 본인의 식당인지 확인
+        if (userRole.equals("점주") && !restaurant.getUserId().equals(userId)) {
+            throw new UnauthorizedAccessException("본인의 식당만 삭제할 수 있습니다.");
+        }
+
         // 연관된 운영시간 논리 삭제
         restaurant.getOperatingHourEntityList().forEach(hour -> hour.deleteOperatingHourEntity(PassportUtil.getUsername(passport)));
